Split scalable marker code into header and source file.
